Updated test structure to support several DB adapters
